Fix bug for compiling Tiktoken
The package Tiktoken, defined in requirements.txt, needs to be build
using a newer version of the rust compiler than the one provided by the
python:3 image.

The solution was to download and install the newest version or Rust.
More information see this thread: https://github.com/openai/tiktoken/issues/23